move: class 위치 변경
