+++ src/main/java/study/spring/solid/part1/application/service/kj/calculator/KJ_Calculator.java
@@ -28,62 +28,3 @@ public class KJ_Calculator {
         return strategy.calculate(op1, op2);
     }
 }
-
-@Component
-class KJ_Add implements KJ_CalculateStrategy {
-
-    @Override
-    public long calculate(long op1, long op2) {
-        return op1 + op2;
-    }
-
-    @Override
-    public CalculateOperation getOperator() {
-        return CalculateOperation.ADD;
-    }
-}
-
-@Component
-class KJ_Sub implements KJ_CalculateStrategy {
-
-    @Override
-    public long calculate(long op1, long op2) {
-        return op1 - op2;
-    }
-
-    @Override
-    public CalculateOperation getOperator() {
-        return CalculateOperation.SUB;
-    }
-}
-
-@Component
-class KJ_Mul implements KJ_CalculateStrategy {
-
-    @Override
-    public long calculate(long op1, long op2) {
-        return op1 * op2;
-    }
-
-    @Override
-    public CalculateOperation getOperator() {
-        return CalculateOperation.MUL;
-    }
-}
-
-@Component
-class KJ_Div implements KJ_CalculateStrategy {
-
-    @Override
-    public long calculate(long op1, long op2) {
-
-        if (op2 != 0) return op1 / op2;
-
-        throw new ArithmeticException("0으로 나눌 수 없습니다.");
-    }
-
-    @Override
-    public CalculateOperation getOperator() {
-        return CalculateOperation.DIV;
-    }
-}
